pyproject.toml
[project]
name = "stt-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "quarriedstone",email = "quarriedstone@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python-telegram-bot = "^22.3"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"
